# internal/web/openapi.yaml
openapi: 3.0.3
info:
  title: OpenTrace API
  description: |
    OpenTrace is a self-hosted observability platform for Ruby on Rails applications.
    It provides log ingestion, error tracking, performance monitoring, health checks,
    user journey analysis, and an AI-powered watch system.

    ## Authentication

    The API uses three authentication schemes:

    - **ApiKeyAuth** -- Bearer token for log/metric ingestion endpoints. Configure via `OPENTRACE_API_KEY` or the Settings UI.
    - **SessionAuth** -- Cookie-based session for browser access. Obtain by POSTing to `/login`.
    - **MCPTokenAuth** -- Bearer token for MCP (Model Context Protocol) connections. Each user has a unique MCP token.
  version: "1.0.0"
  contact:
    name: OpenTrace
    url: https://github.com/adham90/opentrace
servers:
  - url: /
...
security: []
paths:
...
  /login:
    post:
      tags: [Auth]
      summary: Log in
      description: Authenticate with email and password. Sets a session cookie on success.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                remember:
                  type: string
                  description: Set to "on" for a long-lived session
      responses:
        "302":
          description: Redirect to dashboard on success or back to login on failure
        "429":
          description: Rate limited
  /register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                display_name:
                  type: string
      responses:
        "302":
          description: Redirect to dashboard on success
...
  /api/version/check:
    get:
      tags: [Public]
      summary: Check for updates
      parameters:
        - name: force
          in: query
          schema:
            type: string
            enum: ["1"]
          description: Set to "1" to bypass cache
      responses:
        "200":
          description: Update check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  current_version:
                    type: string
                  update_available:
                    type: boolean
                  latest_version:
                    type: string
                  release_url:
                    type: string
...
  /api/logs:
    post:
      tags: [Ingestion]
      summary: Ingest logs
      description: |
        Accepts a single log object or an array of log objects.
        Supports batch deduplication via the `X-Batch-ID` header.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Batch-ID
          in: header
          schema:
            type: string
          description: Unique batch identifier for deduplication
        - name: X-API-Version
          in: header
          schema:
            type: integer
          description: Client API version for compatibility checks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/LogIngest"
                - type: array
                  items:
                    $ref: "#/components/schemas/LogIngest"
      responses:
        "201":
          description: Logs ingested successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
        "200":
          description: Batch deduplicated (already received)
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  deduplicated:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/servers/register:
    post:
      tags: [Ingestion]
      summary: Register a server
      description: Register or update a server for metric collection.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterServerParams"
      responses:
        "200":
          description: Server registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  server_id:
                    type: string
                    format: uuid
                  status:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/services:
    get:
      tags: [Logs]
      summary: List distinct services
      security:
        - SessionAuth: []
      responses:
        "200":
          description: List of service names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /api/event_types:
    get:
      tags: [Logs]
      summary: List distinct event types
      security:
        - SessionAuth: []
      responses:
        "200":
          description: List of event type names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /api/overview:
    get:
      tags: [Logs]
      summary: Dashboard overview
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Overview counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: integer
                  connectors:
                    type: integer
                  servers:
                    type: integer
  /api/tools:
    get:
      tags: [Logs]
      summary: List available MCP tools
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Tool catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    description:
                      type: string
                    access:
                      type: string
                    requires:
                      type: string
  /api/logs/poll:
    get:
      tags: [Logs]
      summary: Poll for new logs
      description: Long-poll endpoint for real-time log updates. Returns HTML fragments or empty 200.
      security:
        - SessionAuth: []
      parameters:
        - name: since_id
          in: query
          schema:
            type: integer
            format: int64
          description: Return logs newer than this ID
      responses:
        "200":
          description: New log entries (HTML) or empty if no new logs
  /api/logs/{id}:
    get:
      tags: [Logs]
      summary: Get log entry detail
      security:
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogEntry"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Connectors ────────────────────────────────────────────────────────
  /api/connectors:
    get:
      tags: [Connectors]
      summary: List connectors
      security:
        - SessionAuth: []
      responses:
        "200":
          description: List of data sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataSource"
    post:
      tags: [Connectors]
      summary: Create a connector
      description: Admin only.
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDataSourceParams"
      responses:
        "201":
          description: Connector created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/connectors/{id}:
    get:
      tags: [Connectors]
      summary: Get connector by ID
      security:
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ConnectorID"
      responses:
        "200":
          description: Connector details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSource"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Connectors]
      summary: Update a connector
      description: Admin only.
      security:
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ConnectorID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDataSourceParams"
      responses:
        "200":
          description: Connector updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSource"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Connectors]
      summary: Delete a connector
      description: Admin only.
      security:
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ConnectorID"
      responses:
        "204":
          description: Connector deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/connectors/{id}/test:
    post:
      tags: [Connectors]
      summary: Test connector connectivity
      description: Admin only.
      security:
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ConnectorID"
      responses:
        "200":
          description: Connector test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSource"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Servers ───────────────────────────────────────────────────────────
...
  /api/watches/alerts/count:
    get:
      tags: [Watches]
      summary: Count active alerts
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Alert count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: API key for log/metric ingestion. Set via OPENTRACE_API_KEY env var or Settings UI.
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Session cookie obtained by logging in via POST /login.
    MCPTokenAuth:
      type: http
      scheme: bearer
      description: Per-user MCP token for Model Context Protocol connections.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...